# docs/controlplane-openapi.yaml
openapi: 3.0.1
info:
  title: Control Plane
  version: 1.0.0
paths:
...
  /cp/routes/{id}:
    put:
      summary: Update route
      parameters:
        - name: If-Match
          in: header
          schema: { type: string }
      responses:
        '204': { description: Updated, headers: { ETag: { description: New tag, schema: { type: string } } } }
        '409': { description: ETag mismatch }
    delete:
      summary: Delete route
      parameters:
        - name: If-Match
          in: header
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '409': { description: ETag mismatch }
...
  /cp/waf/{rule}:
    put:
      summary: Update WAF toggle
      parameters:
        - name: If-Match
          in: header
          schema: { type: string }
      responses:
        '204': { description: Updated, headers: { ETag: { schema: { type: string } } } }
        '409': { description: ETag mismatch }
    delete:
      summary: Delete WAF toggle
      parameters:
        - name: If-Match
          in: header
          schema: { type: string }
      responses:
        '204': { description: Deleted }
        '409': { description: ETag mismatch }
...